pyproject.toml
[tool.poetry]
name = "f1 schedule telegram bot"
version = "0.1.0"
description = "Telegram bot to notify telegram users, groups and channels of upcoming F1 events."
authors = ["Jurriaan Den Toonder <jurriaan@toonder.dev>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "f1_schedule_telegram_bot" },
]
repository = "https://github.com/Fastjur/F1ScheduleTelegramBot"

[tool.poetry.dependencies]
python = "^3.10.6"
python-dotenv = "^1.0.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.2"}
requests = "^2.29.0"
ics = "^0.7.2"
arrow = "^1.2.3"
ergast-py = "^0.7.0"
prettytable = "^3.7.0"


[tool.poetry.group.dev.dependencies]
pylama = {extras = ["toml", "mypy", "pylint", "eradicate", "radon", "vulture"], version = "^8.4.1"}
black = "^23.3.0"
types-requests = "^2.30.0.0"
# Lock pydocstyle to a version lower than 6.2, otherwise the build errors
# https://github.com/klen/pylama/issues/232
pydocstyle = "^6.0.0, <6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "f1_schedule_telegram_bot.main:main"

[tool.poetry.urls]
issues = "https://github.com/Fastjur/F1ScheduleTelegramBot/issues"

[tool.pylama]
linters = "pycodestyle,pydocstyle,pyflakes,mccabe,pylint,radon,mypy,vulture"
ignore = "D202,D203,D212,W1203"
# Skip .venv in github actions
skip = ".venv/*"

[tool.vulture]
min_confidence = 70
